web3_sha3 integration test moved to sc_evm_rpc_web3_methods.py
