+++ qa/sc_evm_rpc_web3_methods.py
@@ -2,7 +2,7 @@
 import re
 
 from SidechainTestFramework.account.ac_chain_setup import AccountChainSetup
-from SidechainTestFramework.scutil import assert_true
+from SidechainTestFramework.scutil import assert_true, assert_false, assert_equal
 
 """
 Check web3 namespace rpc methods
@@ -13,6 +13,7 @@ Configuration:
 
 Test:
     - Check web3_clientVersion
+    - Check web3_sha3
     
 """
 
@@ -21,6 +22,13 @@ class SCEvmRpcWeb3Methods(AccountChainSetup):
 
     def __init__(self):
         super().__init__(number_of_sidechain_nodes=1, withdrawalEpochLength=10)
+
+    def validate(self, response, expected=None):
+        if expected is not None:
+            assert_false("error" in response)
+            assert_equal(expected, response["result"], "unexpected response")
+        else:
+            assert_true("error" in response, response['error'])
 
     def run_test(self):
         sc_node_1 = self.sc_nodes[0]
@@ -35,6 +43,16 @@ class SCEvmRpcWeb3Methods(AccountChainSetup):
         assert_true(len(clientVersion[2]) > 0 and clientVersion[2] != 'dev')
         # Check that jdk version has at least 2 digits
         assert_true(re.search(r'jdk\d{2,}', clientVersion[3]))
+
+        # Check web3_sha3 method
+        self.validate(sc_node_1.rpc_web3_sha3("0x68656c6c6f20776f726c64"),
+                      "0x47173285a8d7341e5e972fc677286384f802f8ef42a5ec5f03bbfa254cb01fad")
+        self.validate(sc_node_1.rpc_web3_sha3("0x"), "0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470")
+        self.validate(sc_node_1.rpc_web3_sha3("0x68656c6c6f20776f726c642"))
+        self.validate(sc_node_1.rpc_web3_sha3("0x68656c6c6f20776f726c6w"))
+        self.validate(sc_node_1.rpc_web3_sha3("68656c6c6f20776f726c6w"))
+        self.validate(sc_node_1.rpc_web3_sha3("68656c6c6f20776f726c64"))
+        self.validate(sc_node_1.rpc_web3_sha3(""))
 
 
 if __name__ == "__main__":
